+++ @types/global.d.ts
@@ -1,0 +1,2 @@
+/** @see https://tsdx.io/optimization#code-dev-code */
+// declare var __DEV__: boolean; // defined in global scope from @types/react-native coincidentally 🤷‍♂️